requirements.txt
Flask>=3.0.0,<4
Werkzeug>=3.0.0,<4
moviepy>=1.0.3,<2.0.0
imageio-ffmpeg>=0.4.9
requests>=2.31.0
gTTS>=2.5.1
google-api-python-client>=2.126.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
selenium>=4.22.0
playwright>=1.45.0
python-dotenv>=1.0.1
celery>=5.4.0
redis>=5.0.4
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.25
sentry-sdk[flask,celery]>=1.40.0
